refactor: fsd layer 위반 코드 제거 (type WishesInfo shared 로 이동)
