+++ packages/client/src/entities/semester/model/useSemesterNameBySubjectId.ts
@@ -1,6 +1,7 @@
 import { SEMESTERS } from '../api/semester';
 
-const semesters = [
+// Todo: 다음 학기에 SPRING_26 주석 해제 할 것
+const SEMESTER_RANGES = [
   { code: 'SPRING_25', startId: 1, endId: 2422 },
   { code: 'SUMMER_25', startId: 2562, endId: 2647 },
   { code: 'FALL_25', startId: 2648, endId: 5144 },
@@ -9,7 +10,7 @@ const semesters = [
 ];
 
 export function useSemesterNameBySubjectId(subjectId: number) {
-  const semester = semesters.find(sem => subjectId >= sem.startId && subjectId <= sem.endId);
+  const semester = SEMESTER_RANGES.find(sem => subjectId >= sem.startId && subjectId <= sem.endId);
 
   const semesterData = SEMESTERS.find(s => s.semesterCode === semester?.code);
