+++ packages/client/src/features/wish/model/useWishesInfo.ts
@@ -1,12 +1,6 @@
 import { useParams } from 'react-router-dom';
 import { useSemesterNameBySubjectId } from '@/entities/semester/model/useSemesterNameBySubjectId';
-
-export interface WishesInfo {
-  subjectId: number;
-  semesterCode: string | undefined;
-  semesterValue: string | undefined;
-  isLastSemesterWish: boolean;
-}
+import { WishesInfo } from '@/shared/model/types';
 
 // 관심과목 상세 페이지에서 subjectId와 semesterCode를 가져오는 커스텀 훅
 // 다른 훅을 사용할 때, 정보를 주입
