+++ packages/client/src/shared/model/types.ts
@@ -67,6 +67,13 @@ export interface Subject {
   // departmentName: string;
 }
 
+export interface WishesInfo {
+  subjectId: number;
+  semesterCode: string | undefined;
+  semesterValue: string | undefined;
+  isLastSemesterWish: boolean;
+}
+
 /** before: 튜토리얼 + 관심과목 선택
  * start: 시작 -> 대기 -> 과목 불러오기 전
  * progress: 과목 신청 프로세스
